+++ lib/ts/recipe/thirdparty/api/implementation.ts
@@ -182,7 +182,6 @@ export default function getAPIInterface(): APIInterface {
                 response.user.id,
                 {},
                 {},
-                undefined,
                 userContext
             );
             return {